correct names of files
