+++ app_temp.py
@@ -1,0 +1,11 @@
+import dash
+import dash_bootstrap_components as dbc
+
+# bootstrap theme
+# https://bootswatch.com/lux/
+external_stylesheets = [dbc.themes.LUX]
+
+app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
+
+server = app.server
+app.config.suppress_callback_exceptions = True
