--- index.py
@@ -1,105 +1,0 @@
-import dash_core_components as dcc
-import dash_html_components as html
-from dash.dependencies import Input, Output, State
-import dash_bootstrap_components as dbc
-
-from app import app
-# import all pages in the app
-from apps import home, about, research, components, showcase, forum, contribute, resources, outreach
-
-# building the navigation bar
-# https://github.com/facultyai/dash-bootstrap-components/blob/master/examples/advanced-component-usage/Navbars.py
-dropdown = dbc.DropdownMenu(
-    children=[
-        dbc.DropdownMenuItem("Home", href="/home"),
-        dbc.DropdownMenuItem("About", href="/about"),
-        dbc.DropdownMenuItem("Research", href="/research"),
-        dbc.DropdownMenuItem("Components", href="/components"),
-        dbc.DropdownMenuItem("Showcase", href="/showcase"),
-        dbc.DropdownMenuItem("Forum", href="/forum"),
-        dbc.DropdownMenuItem("Contribute", href="/contribute"),
-        dbc.DropdownMenuItem("Resources for Educators", href="/resources"),
-        dbc.DropdownMenuItem("Outreach", href="/outreach"),
-    ],
-    nav = True,
-    in_navbar = True,
-    label = "Explore",
-)
-
-navbar = dbc.Navbar(
-    dbc.Container(
-        [
-            html.A(
-                # Use row and col to control vertical alignment of logo / brand
-                dbc.Row(
-                    [
-                        dbc.Col(html.Img(src="/assets/logo.png", height="100px")),
-                        dbc.Col(dbc.NavbarBrand("Robot Muscle Toolkit", className="ml-2")),
-                    ],
-                    align="center",
-                    no_gutters=True,
-                ),
-                href="/home",
-            ),
-            dbc.NavbarToggler(id="navbar-toggler2"),
-            dbc.Collapse(
-                dbc.Nav(
-                    # right align dropdown menu with ml-auto className
-                    [dropdown], className="ml-auto", navbar=True
-                ),
-                id="navbar-collapse2",
-                navbar=True,
-            ),
-        ]
-    ),
-    color="light",
-    dark=False,
-    className="mb-4",
-)
-
-def toggle_navbar_collapse(n, is_open):
-    if n:
-        return not is_open
-    return is_open
-
-for i in [2]:
-    app.callback(
-        Output(f"navbar-collapse{i}", "is_open"),
-        [Input(f"navbar-toggler{i}", "n_clicks")],
-        [State(f"navbar-collapse{i}", "is_open")],
-    )(toggle_navbar_collapse)
-
-# embedding the navigation bar
-app.layout = html.Div([
-    dcc.Location(id='url', refresh=False),
-    navbar,
-    html.Div(id='page-content')
-])
-
-
-@app.callback(Output('page-content', 'children'),
-              [Input('url', 'pathname')])
-def display_page(pathname):
-    if pathname == '/home':
-        return home.layout
-    elif pathname == '/about':
-        return about.layout
-    elif pathname == '/research':
-        return research.layout
-    elif pathname == '/components':
-        return components.layout
-    elif pathname == '/showcase':
-        return showcase.layout
-    elif pathname == '/forum':
-        return forum.layout
-    elif pathname == '/contribute':
-        return contribute.layout
-    elif pathname == '/resources':
-        return resources.layout
-    elif pathname == '/outreach':
-        return outreach.layout
-    else:
-        return home.layout
-
-if __name__ == '__main__':
-    app.run_server(host='127.0.0.1', debug=True)
